Restructure channel folder
