+++ lib/screens/channel/channel.dart
@@ -12,12 +12,12 @@ import 'package:frosty/constants/constants.dart';
 import 'package:frosty/core/auth/auth_store.dart';
 import 'package:frosty/main.dart';
 import 'package:frosty/screens/channel/chat/chat.dart';
-import 'package:frosty/screens/channel/stores/chat_assets_store.dart';
-import 'package:frosty/screens/channel/stores/chat_details_store.dart';
-import 'package:frosty/screens/channel/stores/chat_store.dart';
-import 'package:frosty/screens/channel/stores/video_store.dart';
-import 'package:frosty/screens/channel/widgets/video.dart';
-import 'package:frosty/screens/channel/widgets/video_overlay.dart';
+import 'package:frosty/screens/channel/chat/details/chat_details_store.dart';
+import 'package:frosty/screens/channel/chat/stores/chat_assets_store.dart';
+import 'package:frosty/screens/channel/chat/stores/chat_store.dart';
+import 'package:frosty/screens/channel/video/video.dart';
+import 'package:frosty/screens/channel/video/video_overlay.dart';
+import 'package:frosty/screens/channel/video/video_store.dart';
 import 'package:frosty/screens/settings/settings.dart';
 import 'package:frosty/screens/settings/settings_store.dart';
 import 'package:frosty/widgets/modal.dart';
