+++ lib/screens/channel/chat/details/chat_details.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
+import 'package:frosty/screens/channel/chat/details/chat_details_store.dart';
 import 'package:frosty/screens/channel/chat/details/chat_modes.dart';
 import 'package:frosty/screens/channel/chat/details/chat_users_list.dart';
-import 'package:frosty/screens/channel/stores/chat_details_store.dart';
-import 'package:frosty/screens/channel/stores/chat_store.dart';
+import 'package:frosty/screens/channel/chat/stores/chat_store.dart';
 import 'package:frosty/widgets/modal.dart';
 
 class ChatDetails extends StatefulWidget {
