+++ lib/screens/channel/chat/details/chat_details_store.dart
