+++ lib/screens/channel/chat/emote_menu/emote_menu_panel.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:frosty/models/emotes.dart';
 import 'package:frosty/screens/channel/chat/emote_menu/emote_menu_section.dart';
-import 'package:frosty/screens/channel/stores/chat_store.dart';
+import 'package:frosty/screens/channel/chat/stores/chat_store.dart';
 import 'package:frosty/widgets/section_header.dart';
 
 class EmoteMenuPanel extends StatelessWidget {
