+++ lib/screens/channel/chat/stores/chat_store.dart
@@ -6,8 +6,8 @@ import 'package:frosty/core/auth/auth_store.dart';
 import 'package:frosty/models/badges.dart';
 import 'package:frosty/models/emotes.dart';
 import 'package:frosty/models/irc.dart';
-import 'package:frosty/screens/channel/stores/chat_assets_store.dart';
-import 'package:frosty/screens/channel/stores/chat_details_store.dart';
+import 'package:frosty/screens/channel/chat/details/chat_details_store.dart';
+import 'package:frosty/screens/channel/chat/stores/chat_assets_store.dart';
 import 'package:frosty/screens/settings/settings_store.dart';
 import 'package:mobx/mobx.dart';
 import 'package:wakelock/wakelock.dart';
