+++ lib/screens/channel/video/video_overlay.dart
@@ -5,7 +5,7 @@ import 'package:flutter/services.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:frosty/constants/constants.dart';
 import 'package:frosty/main.dart';
-import 'package:frosty/screens/channel/stores/video_store.dart';
+import 'package:frosty/screens/channel/video/video_store.dart';
 import 'package:frosty/widgets/button.dart';
 import 'package:frosty/widgets/dialog.dart';
 import 'package:frosty/widgets/profile_picture.dart';
